General code cleanup.
 - Fixed some warnings in the pool memory implementation to store the
memory addresses as a void*.

 - Added gitignore entry for the vscode folder in case I switch to
vscode eventually.

+++ memory.c
@@ -132,7 +132,7 @@ void kmKernelFree(void* ptr)
         }
     }
     
-    kWriteLog_format1d("Could not find allocation for %d", ptr);
+    kWriteLog_format1d("Could not find allocation for %d", (uint32_t)ptr);
 }
 
 void kmKernelCopy(void* ptrFrom, void* ptrTo)
